requirements.txt
python-telegram-bot==20.8
APScheduler==3.10.4
python-dotenv==1.0.0
luma.oled==3.12.0
pillow==10.0.1
